Fix broken creature doors
