+++ src/open_samus_returns_rando/samus_returns_patcher.py
@@ -6,8 +6,6 @@ from pathlib import Path
 from mercury_engine_data_structures.file_tree_editor import OutputFormat
 
 from open_samus_returns_rando.debug import debug_custom_pickups, debug_spawn_points
-from open_samus_returns_rando.doors.custom_doors import patch_custom_doors
-from open_samus_returns_rando.doors.door_patches import patch_doors
 from open_samus_returns_rando.files import files_path
 from open_samus_returns_rando.logger import LOG
 from open_samus_returns_rando.lua_editor import LuaEditor
@@ -21,6 +19,7 @@ from open_samus_returns_rando.pickups.custom_pickups import patch_custom_pickups
 from open_samus_returns_rando.pickups.pickup import patch_pickups
 from open_samus_returns_rando.specific_patches import cosmetic_patches, game_patches, tunable_patches
 from open_samus_returns_rando.specific_patches.chozo_seal_patches import patch_chozo_seals
+from open_samus_returns_rando.specific_patches.door_patches import patch_doors
 from open_samus_returns_rando.specific_patches.heat_room_patches import patch_heat_rooms
 from open_samus_returns_rando.specific_patches.hint_patches import patch_hints
 from open_samus_returns_rando.specific_patches.map_icons import patch_tiles
@@ -84,8 +83,7 @@ def patch_extracted(input_path: Path, output_path: Path, configuration: dict) ->
     patch_heat_rooms(editor)
 
     # Patch door types and make shields on both sides
-    patch_custom_doors(editor, configuration["custom_doors"])
-    patch_doors(editor, configuration["door_patches"])
+    patch_doors(editor, configuration["door_patches"], configuration["custom_doors"])
 
     # Patch tunables
     tunable_patches.patch_tunables(editor, configuration.get("reserves_per_tank", {}))
